Fix random at each step, fix logprob ratio, increase steps/episode to 300.
Previously, we were getting a total reward of 0 at every epoch.
Fix randomization (each step was identical in previous code) and increase steps/episode.
Change entropy loss to be - in total loss to not have a negative in LossInformation object.
Increase LR as average reward is like 40 / 20 episodes for 128 envs and 200 steps.
Increase steps further to guarantee non-zero gradients (although the rewards are super small at that point).